Refactor user deletion
